moved carousel to components
